+++ user/src/main/java/com/zerobase/user/point/domain/model/PointHistoryEntity.java
@@ -1,5 +1,6 @@
 package com.zerobase.user.point.domain.model;
 
+import com.zerobase.user.point.domain.model.payment.PointPaymentOrderEntity;
 import jakarta.persistence.*;
 import lombok.*;
