+++ user/src/main/java/com/zerobase/user/point/domain/model/payment/PointPaymentOrderEntity.java
@@ -1,5 +1,7 @@
-package com.zerobase.user.point.domain.model;
+package com.zerobase.user.point.domain.model.payment;
 
+import com.zerobase.user.point.domain.model.PaymentMethod;
+import com.zerobase.user.point.domain.model.PaymentStatus;
 import jakarta.persistence.*;
 import lombok.*;
