+++ user/src/main/java/com/zerobase/user/point/domain/repository/PointStoreImpl.java
@@ -1,8 +1,5 @@
 package com.zerobase.user.point.domain.repository;
 
-import com.zerobase.user.point.domain.model.PointPaymentOrderEntity;
-import com.zerobase.user.point.domain.repository.payment.KakaoPaymentTransactionRepository;
-import com.zerobase.user.point.domain.repository.payment.TossPaymentTransactionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
