+++ user/src/main/java/com/zerobase/user/point/domain/repository/payment/PaymentStore.java
@@ -1,6 +1,6 @@
 package com.zerobase.user.point.domain.repository.payment;
 
-import com.zerobase.user.point.domain.model.PointPaymentOrderEntity;
+import com.zerobase.user.point.domain.model.payment.PointPaymentOrderEntity;
 import com.zerobase.user.point.domain.model.payment.PaymentTransactionEntity;
 
 public interface PaymentStore {
