+++ user/src/main/java/com/zerobase/user/point/domain/repository/payment/PointPaymentOrderRepository.java
@@ -1,6 +1,6 @@
-package com.zerobase.user.point.domain.repository;
+package com.zerobase.user.point.domain.repository.payment;
 
-import com.zerobase.user.point.domain.model.PointPaymentOrderEntity;
+import com.zerobase.user.point.domain.model.payment.PointPaymentOrderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PointPaymentOrderRepository extends JpaRepository<PointPaymentOrderEntity, Long> {
