+++ user/src/main/java/com/zerobase/user/point/service/payment/KakaoPayment.java
@@ -1,7 +1,7 @@
 package com.zerobase.user.point.service.payment;
 
 import com.zerobase.user.point.domain.model.PaymentMethod;
-import com.zerobase.user.point.domain.model.PointPaymentOrderEntity;
+import com.zerobase.user.point.domain.model.payment.PointPaymentOrderEntity;
 import com.zerobase.user.point.domain.model.payment.KakaoPaymentTransactionEntity;
 import com.zerobase.user.point.domain.model.payment.PayStatus;
 import com.zerobase.user.point.domain.model.payment.PaymentTransactionEntity;
