+++ user/src/main/java/com/zerobase/user/point/service/payment/Payment.java
@@ -1,7 +1,7 @@
 package com.zerobase.user.point.service.payment;
 
 import com.zerobase.user.point.domain.model.PaymentMethod;
-import com.zerobase.user.point.domain.model.PointPaymentOrderEntity;
+import com.zerobase.user.point.domain.model.payment.PointPaymentOrderEntity;
 import com.zerobase.user.point.domain.model.payment.PaymentTransactionEntity;
 
 public interface Payment {
