+++ user/src/main/java/com/zerobase/user/point/service/payment/PaymentEntityFactory.java
@@ -2,7 +2,7 @@ package com.zerobase.user.point.service.payment;
 
 import com.zerobase.user.exception.BaseException;
 import com.zerobase.user.exception.PaymentErrorCode;
-import com.zerobase.user.point.domain.model.PointPaymentOrderEntity;
+import com.zerobase.user.point.domain.model.payment.PointPaymentOrderEntity;
 import com.zerobase.user.point.domain.model.payment.PaymentTransactionEntity;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
