+++ user/src/main/java/com/zerobase/user/security/TokenProvider.java
@@ -1,7 +1,6 @@
 package com.zerobase.user.security;
 
 import com.zerobase.user.application.UserFacade;
-import com.zerobase.user.exception.BaseException;
 import com.zerobase.user.exception.BasicErrorCode;
 import com.zerobase.user.exception.JwtCustomException;
 import io.jsonwebtoken.Claims;
@@ -38,12 +37,12 @@ public class TokenProvider {
         claims.put(KEY_ROLES, userSecurity.getUserRoles());
 
         return Jwts.builder()
-                .subject(userSecurity.getUserUuid())
-                .claims(claims)
-                .issuedAt(now) // 토큰 생성 시간
-                .expiration(expiredDate) // 토큰 만료 시간
-                .signWith(secretKey) // 사용할 암호화 알고리즘, 비밀키
-                .compact();
+            .subject(userSecurity.getUserUuid())
+            .claims(claims)
+            .issuedAt(now) // 토큰 생성 시간
+            .expiration(expiredDate) // 토큰 만료 시간
+            .signWith(secretKey) // 사용할 암호화 알고리즘, 비밀키
+            .compact();
 
     }
 
@@ -69,10 +68,10 @@ public class TokenProvider {
 
         try {
             return Jwts.parser()
-                    .verifyWith(secretKey)
-                    .build()
-                    .parseSignedClaims(token)
-                    .getPayload();
+                .verifyWith(secretKey)
+                .build()
+                .parseSignedClaims(token)
+                .getPayload();
 
         } catch (ExpiredJwtException e) {
             throw new JwtCustomException(BasicErrorCode.EXPIRED_TOKEN);
